+++ kermit/k95/ckocon.c
@@ -793,17 +793,80 @@ clearcmdscreen(void) {
 /*---------------------------------------------------------------------------*/
 /* clearscrollback                                          | Page: First    */
 /*---------------------------------------------------------------------------*/
-/* Clears the scrollback, which is associated with the first page only       */
+/* Clears the scrollback, which is associated with the first page only.      */
+/* Optionally preserves the current contents of the screen, erasing          */
+/* scrollback only. When preserving the current screen, scrollback data is   */
+/* not erased from memory, it is just rendered inaccessible and will         */
+/* eventually be overwritten by new data.                                    */
 void
-clearscrollback( BYTE vmode ) {
-    ULONG bufsize = VscrnGetPageBufferSize(vmode, FALSE, 0) ;
+clearscrollback( BYTE vmode, BOOL keep_screen ) {
+    int page = 0; /* Only page 0 has scrollback */
+    int len;
 
+    RequestVscrnMutex( vmode, SEM_INDEFINITE_WAIT ) ;
+
+    if (keep_screen) {
+        int i;
+
+        /* We can't get away with just setting the page top as the beginning of
+         * the scrollback - KUI won't show this as "scrollback empty", and
+         * VscrnScrollPage will make the old scrollback available again on the
+         * next scroll. So instead we copy the current screen to the top of the
+         * vscreen buffer, then reset the beginning, top and end of the page.
+         */
+
+        VscrnGetPageTop(vmode, FALSE, page) ;
+        len = VscrnGetHeight(vmode) -(tt_status[vmode]?2:1);
+
+        /* Copy the contents of the line to the top of the vscreen buffer */
+        for (i = 0; i <= len; i++) {
+            VscrnCopyLine(VscrnGetPageLineFromTop(vmode, i, page),
+                VscrnGetPageLine(vmode, i, page));
+        }
+
+        VscrnSetPageTop(vmode, 0, FALSE, page, TRUE);
+        VscrnSetPageBegin(vmode, 0, page);
+        VscrnSetPageEnd(vmode, len, page) ;
+
+        scrollstatus[vmode] = FALSE ;
+        scrollflag[vmode] = FALSE ;
+        cursoron[vmode] = FALSE ;
+
+        if ( IsConnectMode() || vmode != VTERM )
+            VscrnIsDirty(vmode);
+
+        return;
+    }
+
+#ifdef COMMENT
+    /* Prior to 3.0 beta 8 this is how K95 always cleared the scrollback,
+     * though I'm not sure why as it's more work than just moving the top,
+     * begin and end. It's also not safe given K95s multi-threaded nature
+     * as the RdComWrtScr thread isn't great about ensuring it has the vscrn
+     * mutex before doing things. Because of this, the K_CLRSCROLL can
+     * cause a crash if the RdComWrtScr thread is busy doing stuff with
+     * the vscrn.
+     */
+    ULONG bufsize = VscrnGetPageBufferSize(vmode, FALSE, 0) ;
     VscrnSetBufferSize( vmode, 256, vscrn[vmode].page_count ) ;
     VscrnSetBufferSize( vmode, bufsize, vscrn[vmode].page_count ) ;
+#else
+    /* Instead, just reuse the existing buffer. VscrnScrollPage erases lines as
+     * it reuses them, so the old stuff in the buffer won't ever be visible.
+     * This is much safer and *seems* to work just as well. */
+    VscrnGetPageTop(vmode, FALSE, page) ;
+    len = VscrnGetHeight(vmode) -(tt_status[vmode]?2:1);
+    VscrnSetPageTop(vmode, 0, FALSE, page, TRUE);
+    VscrnSetPageBegin(vmode, 0, page);
+    VscrnSetPageEnd(vmode, len, page) ;
+#endif /* COMMENT */
+
     scrollstatus[vmode] = FALSE ;
     scrollflag[vmode] = FALSE ;
     cursoron[vmode] = FALSE ;
     cleartermpage(vmode, 0) ;
+
+    ReleaseVscrnMutex(vmode);
 }
 
 /*---------------------------------------------------------------------------*/